+++ bin/typedef/linetv-clip-data.js
@@ -1,0 +1,12 @@
+/**
+ * @typedef LINETvClipData
+ * @property {number} clipNo
+ * @property {string} clipTitle
+ * @property {number} likeitPoint
+ * @property {number} playCount
+ * @property {string} displayPlayTime
+ * @property {string} thumbnailUrl
+ * @property {string} serviceUrl
+ */
+"use strict";
+//# sourceMappingURL=linetv-clip-data.js.map