+++ bin/typedef/linetv-get-category-response-data.js
@@ -1,0 +1,7 @@
+/**
+ * @typedef LINETvGetCategoryResponseData
+ * @property {{ message: string; code: number }} header
+ * @property {{ representClip: LINETvClipData; tabs: Array<LINETvChannelData>; hasMore: boolean }} body
+ */
+"use strict";
+//# sourceMappingURL=linetv-get-category-response-data.js.map