+++ bin/typedef/linetv-get-spotlight-response-data.js
@@ -1,0 +1,7 @@
+/**
+ * @typedef LINETvGetSpotlightResponseData
+ * @property {{ message: string; code: number }} header
+ * @property {{ clips: Array<LINETvClipData>|null; channels: Array<LINETvChannelData>|null; playlists: Array<LINETvPlaylistData>|nullZ; moduleName: string }} body
+ */
+"use strict";
+//# sourceMappingURL=linetv-get-spotlight-response-data.js.map