+++ bin/typedef/linetv-search-response-data.js
@@ -1,0 +1,7 @@
+/**
+ * @typedef LINETvSearchResponseData
+ * @property {{ message: string; code: number }} header
+ * @property {{ clips: Array<LINETvClipData> ; hasMore: boolean; total: number}} body
+ */
+"use strict";
+//# sourceMappingURL=linetv-search-response-data.js.map