+++ bin/typedef/richmenu-list-response-data.js
@@ -1,6 +1,6 @@
 /**
  * @typedef RichmenuListResponseData
- * @property {Array<RichMenuData>} richmenus:
+ * @property {Array<RichMenuData>} richmenus
  */
 "use strict";
 //# sourceMappingURL=richmenu-list-response-data.js.map