+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [1.3.x]
+### Added
+ - `linetv list:modules` Gets a spotlight data unit list such as represent_clip, editor_pick_clip, and theme_playlist. 
+ - `linetv get:spotlight` Gets spotlight data.
+ - `linetv get:spotlight --format json` Gets spotlight data in JSON format.
+ - `linetv list:category` Gets a category list such as drama, music, etc.
+ - `linetv get:category` Gets category home data.
+ - `linetv ranking` Gets clip ranking data.
+ - `linetv search` Gets a clip search result.
+ - `linetv list:station` Gets the station home (TV station) list.
+
+### Fixed
+ - Fixed `npm test` failed on Windows
+ - Fixed missing qs module
+
 ## [1.2.0]
 ### Added
  - `things register:scenario-set` Register (create or update) a scenario set for automatic communication under a product.